Intento de correccion de porcentaje de musicos por evento y metodos de calculo de ingresos casi terminados

+++ CastillaCanoAlvaro-Proyecto1/src/main/java/com/salesianostriana/dam/castillacanoalvaroproyecto1/repositorio/MusicoRepositorio.java
@@ -28,7 +28,7 @@ public interface MusicoRepositorio extends JpaRepository<Musico, Long>{
 			WHERE e.id = :id
 			GROUP BY e.id
 			""")
-	public double findNumeroMusicosPorEvento(@Param("id") long id);
+	public int findNumeroMusicosPorEvento(@Param("id") long id);
